+++ FindPOLine/utils.js
@@ -12,7 +12,10 @@ import {
 } from '@folio/stripes-acq-components';
 
 import { FILTERS, QUALIFIER_SEPARATOR } from './constants';
-import { getCqlQuery, getKeywordQuery } from './OrderLinesSearchConfig';
+import {
+  getCqlQuery,
+  getKeywordQuery,
+} from './OrderLinesSearchConfig';
 
 const defaultSearchFn = (localeDateFormat, customFields = []) => (query, qindex) => {
   if (qindex === 'details.productIds') {